Update Next.js configuration for export output
We want to use Sevalla static website hosting as it is free.
To do this we need to export a static export.

Changed the output setting from 'standalone' to 'export' in the Next.js configuration to facilitate static site generation. Commented out the images optimization settings for potential future use.