🎨 Revamp Contact, Hero, and Products sections for a personal touch
The website was previously designed for the company handsala ab, but as the company is only one person, it would make more sense to expose that person rather that the company.

Updated the content in the Contact, Hero, and Products sections to reflect a more personal and engaging narrative. Changed headings and descriptions to better convey personality and creativity, enhancing user connection. Added direct contact links and playful descriptions of projects to showcase skills and invite collaboration.