🎨 Make background noise same on all pages
The background noise was shown in different sizes for different pages depending on the size of the viewbox of the container it is applied to. But we want the noise to stay the same across pages to give a feeling of unity.

This adds a background-size property to the noise, making it the same size no matter the size of the containers viewbox.